+++ src/java/info/toegepaste/www/service/ResultatenService.java
@@ -26,5 +26,5 @@ public interface ResultatenService {
     public boolean updateScore(Score score);
     
     public List<String> getTotaalVoorVakken(int studentId);
-    public double getPercentageVoorStudent(int studentId);
+    public long getPercentageVoorStudent(int studentId);
 }
